fix(ui): Make sure styles match design system;
After some feedback from Kyle, I've gone and made the styles more
accessible for moving around the application using the keyboard and
making sure that things don't collide with people who use Touch or Mouse
inputs. Inspiration for the focus trick is to leverage a suggestion from
@LeaVerou mentioned here: https://x.com/LeaVerou/status/1045768279753666562

This also updates the colors to match the Figma design which was missed
in the earlier changeset.