fix(ui): Update <ConfigurationsTable /> & Page;
This modifies both the styling and the structure of the
`<ConfigurationsTable />` component. The styles now support mobile
styles that are included with USWDS for Tables. The styling has also
been modified a bit to work on all screen sizes. The Configurations Page
also has some updates to the `Set up new condition` button as well to
have it match the designs more. With this design we lose the ability to
have the columns match the design specifically, but is the easier to
maintain version to have the table spaced out evenly depending on the
contents of the columns.

Co-Authored-By: Jake Wheeler <jake.wheeler@skylight.digital>